Move `LoadingCard` to new file
